Adding command to Remap the cluster to UC
