+++ src/databricks/labs/ucx/cli.py
@@ -21,6 +21,7 @@ from databricks.labs.ucx.hive_metastore.mapping import TableMapping
 from databricks.labs.ucx.hive_metastore.table_migrate import TableMove, TablesMigrate
 from databricks.labs.ucx.install import WorkspaceInstallation
 from databricks.labs.ucx.workspace_access.groups import GroupManager
+from databricks.labs.ucx.workspace_access.clusters import ClusterAccess
 
 ucx = App(__file__)
 logger = get_logger(__file__)
@@ -329,8 +330,8 @@ def cluster_remap(w: WorkspaceClient, cluster_id):
     """Re-mapping the cluster to UC"""
     if not cluster_id:
         raise KeyError("You did not specify --cluster_id")
-    installation = WorkspaceInstallation.current(w)
-    installation.map_cluster_to_uc(cluster_id)
+    cluster = ClusterAccess.current(w)
+    cluster.map_cluster_to_uc(cluster_id)
 
 
 if __name__ == "__main__":
