+++ src/databricks/labs/ucx/install.py
@@ -906,21 +906,6 @@ class WorkspaceInstallation:
         run_id = latest_job_run.run_id
         return job_id, run_id
 
-    def map_cluster_to_uc(self, cluster_id):
-        try:
-            spark_version = self._ws.clusters.select_spark_version(latest=True)
-            security_modes = {"Single User": "SINGLE_USER", "Shared": "USER_ISOLATION"}
-            access_mode = self._prompts.choice_from_dict("Select cluster access mode", security_modes)
-            if access_mode == 'SINGLE_USER':
-                data_security_mode = compute.DataSecurityMode.SINGLE_USER
-            else:
-                data_security_mode = compute.DataSecurityMode.USER_ISOLATION
-            self._ws.clusters.edit(
-                cluster_id=cluster_id, spark_version=spark_version, data_security_mode=data_security_mode
-            )
-        except InvalidParameterValue as e:
-            logger.warning(f"skipping cluster remapping: {e}")
-
     def uninstall(self):
         if self._prompts and not self._prompts.confirm(
             "Do you want to uninstall ucx from the workspace too, this would "
